+++ app/src/main/java/com/study/thesuperiorstanislav/decisiontheorylabs/lab2/domain/usecase/DoTheThingLab2.kt
@@ -2,7 +2,7 @@ package com.study.thesuperiorstanislav.decisiontheorylabs.lab2.domain.usecase
 
 import com.study.thesuperiorstanislav.decisiontheorylabs.UseCase
 import com.study.thesuperiorstanislav.decisiontheorylabs.lab1.domain.model.Point
-import com.study.thesuperiorstanislav.decisiontheorylabs.utils.Regression
+import com.study.thesuperiorstanislav.decisiontheorylabs.utils.Math.Regression
 
 class DoTheThingLab2: UseCase<DoTheThingLab2.RequestValues, DoTheThingLab2.ResponseValue>() {
